server/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
pandas==2.2.1
scikit-learn==1.4.1.post1
numpy==1.26.4
pymongo==4.6.2
motor==3.3.2
python-dotenv==1.0.1
pydantic==2.6.3
apscheduler==3.10.4
requests==2.31.0